// src/app/models/profesor.model.ts
export interface Profesor {
  _id: string;          // Identificador único del profesor
  nombre: string;       // Nombre del profesor
  apellido: string;     // Apellido del profesor
  telefono: string;     // Teléfono del profesor
  email: string;        // Correo electrónico del profesor
  especialidad: string; // Especialidad del profesor (por ejemplo, Matemáticas, Historia)
}